requirements.txt
python3
pillow
django
djangorestframework